# src/swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Chillka Rest API",
    "version": "1.0.0",
    "description": ""
  },
  "basePath": "/",
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "scheme": "bearer",
      "description": "Prefix in `Bearer ${token}`"
    }
  },
  "paths": {
...
    "/api/register": {
      "post": {
        "tags": ["Auth"],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/schemas/UserRegisterCredentials"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "tags": ["Auth"],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/schemas/UserLoginCredentials"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/verify-email": {
      "get": {
        "tags": ["Auth"],
        "description": "",
        "parameters": [
          {
            "name": "validateCode",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": ["Auth"],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/activities/{activityId}": {
      "get": {
        "tags": ["Activity"],
        "description": "",
        "parameters": [
          {
            "name": "activityId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "userId",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}